Allowed for passing classes in by std::weak_ptr or by C-style pointer.
Regardless of the type passed in, the classes act identically inside Lua.
The difficulty lies in getting them out afterward.
The output can be requested as a std::shared_ptr, std::weak_ptr, or C-style pointer.
A shared_ptr can return as a shared_ptr or a C-style pointer.
A weak_ptr can return as a weak_ptr, or as either a shared_ptr or a C-style pointer if the weak_ptr is still valid.
A C-style pointer can only be returned as a C-style pointer.

An untrusted user script may store a parameter in a global variable.
If untrusted scripts are being run,
  it is the C++ programmer's responsibility to ensure that any values passed by C-style pointer
  remain valid until the LuaState closes.