Made reference passing to lua be safe.
Whenever a reference is passed, it has a unique id added to a std::set.
When the reference is returned or has a method called,
  library checks whether its id is inside the set.
If not, either LuaExpiredReference is thrown (returning a reference),
  or nil is returned (method call).

This should never happen, unless a user is passed a reference into a function,
  saves the reference as a global variable,
  then uses it the following time.
This is mostly to prevent against this potentially malicious action.

Closes #16.